+++ test/test_boost.py
@@ -139,6 +139,7 @@ class TestBOOSTERASURE:
 
         cppyy.include("boost/type_erasure/any.hpp")
         cppyy.include("boost/type_erasure/member.hpp")
+        cppyy.include("boost/mpl/vector.hpp")
 
     def test01_erasure_usage(self):
         """boost::type_erasure usage"""